requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
agora-token-builder==1.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.5
requests==2.31.0
gunicorn==21.2.0 